[2024-05-29] controller 정리
