+++ src/main/java/com/ohgiraffers/hitechautoworks/config/SecurityConfig.java
@@ -43,8 +43,8 @@ public class SecurityConfig {
         http.authorizeHttpRequests( auth -> {
             auth.requestMatchers("/main","/","/error/*","/auth/idCheck","/auth/searchForId","/auth/searchForPW","/member/login","/member/regist","/certified/loading","/certified/checkinfo").permitAll();
             auth.requestMatchers("/*").hasAnyAuthority(UserRole.ADMIN.getRole());
-            auth.requestMatchers("/employee/**").hasAnyAuthority(UserRole.EMPLOYEE.getRole());
-            auth.requestMatchers("/user/**").hasAnyAuthority(UserRole.EMPLOYEE.getRole(), UserRole.CUSTOMER.getRole());
+            auth.requestMatchers("/employee/**").hasAnyAuthority(UserRole.EMPLOYEE.getRole(),UserRole.ADMIN.getRole());
+            auth.requestMatchers("/user/**").hasAnyAuthority(UserRole.EMPLOYEE.getRole(), UserRole.CUSTOMER.getRole(),UserRole.ADMIN.getRole());
             auth.requestMatchers("/certified/**").hasAnyAuthority(UserRole.CERTIFIED.getRole());
             auth.anyRequest().authenticated();
         }).formLogin(login -> {
