--- src/main/resources/templates/customer/dashboard.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>일반회원 DASHBOARD</title>
-</head>
-<body>
-    <h1>일반회원 DASHBOARD 입니다.</h1>
-</body>
-</html>
