--- src/main/resources/templates/customer/res/res_06.html
@@ -1,173 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>res_06</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
-
-    <style>
-        .navbar-brand img {
-            width: 80%; /* 이미지의 원래 크기로 설정 */
-        }
-        .navbar-nav .nav-item {
-            margin-right: 100px;
-        }
-        #content-container {
-            display: flex;
-            margin-left: 300px;
-            margin-top: -150px;
-
-        }
-
-        #content {
-            width: 80%; /* 화면 너비의 50%로 조정 */
-            height: 600px; /* 회색 배경의 높이 조정 */
-            background-color: lightgray;
-            padding: 100px;
-            white-space: nowrap;
-        }
-        .form-label {
-            font-weight: bold;
-            text-align: right; /* 라벨을 오른쪽으로 정렬 */
-            padding-right: 10px; /* 오른쪽 여백 추가 */
-        }
-
-
-        .btn-regist {
-            background-color: darkgray;
-            border-color: darkgray;
-            margin-top: 120px;
-
-        }
-        nav{
-            white-space: nowrap;
-        }
-
-        #title{
-            text-align: left;
-            margin-left: 100px;
-
-        }
-
-        .rounded-div {
-            background-color: white;
-            border-radius: 20px; /* 테두리 둥글기 정도를 조절합니다. */
-            padding: 20px; /* 선택 사항: 시각적으로 더 나은 외관을 위해 여백을 추가할 수 있습니다. */
-            margin-left: 20px;
-        }
-
-    </style>
-</head>
-
-<body>
-
-<div class="container-fluid">
-
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <div class="container">
-            <a class="navbar-brand" href="#"><img src="https://github.com/panghunJO/HitechProject1/assets/157236060/ade911fb-5b0b-409d-b99a-4b4d23ecff64" alt="Logo" ></a>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0" >
-                    <li class="nav-item">
-                        <a class="nav-link" href="#"></a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#"></a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#"></a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#"></a>
-                    </li>
-                </ul>
-                <p id="username" class="nav-link"  th:text="|${userName}님|"></p>
-                <ul class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                    </a>
-                    <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                        <li><a class="dropdown-item" href="#">마이페이지</a></li>
-                        <li><a class="dropdown-item" href="#">로그아웃</a></li>
-                    </ul>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
-    <div class="container mt-5">
-        <label class="form-label" style="font-size: 20px;">수리예약</label><br><br>
-        <a class="nav-link" href="#" >수리조회</a><br>
-        <a class="nav-link" href="#" style="font-weight: bold;">수리예약</a><br>
-    </div>
-
-    <div class="row">
-        <div class="col-md-12">
-            <div class="container" id="content-container">
-                <div class="container" id="content">
-
-                    <div class="container">
-                        <div class="row">
-                            <div class="col-md-3">
-                                <div style="display: flex; align-items: center;"> <!-- 라벨과 테이블을 수평 정렬하기 위한 flex 컨테이너 -->
-                                    <label id="title" style="text-align: left; margin-right: 10px;">요청목록</label> <!-- 왼쪽 정렬 및 오른쪽 여백 추가 -->
-                                    <div class="rounded-div">
-                                        <table style="width: 400px; height: 250px">
-                                            <tr th:each="list : ${partlist}">
-                                                <td th:text="${list}"></td>
-                                            </tr>
-                                        </table>
-                                    </div>
-                                </div>
-                                <div class="row justify-content-end ">
-                                    <div class="col" style="margin-right: -250px;">
-                                        <button type="button" class="btn btn-regist" onclick="location.href='res_01'">이전</button>
-                                        <button type="button" class="btn btn-regist" onclick="location.href='res_08'">완료</button>
-                                        <button type="button" class="btn btn-regist" onclick="location.href='res_08'">추가 요청사항</button>
-                                        <button type="button" class="btn btn-regist" onclick="location.href='res_07'">예약일</button>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
-<script>
-    // 사용자 이름 클릭 시 서브메뉴 토글
-    document.addEventListener("DOMContentLoaded", function() {
-        var usernameDiv = document.getElementById('username');
-        var submenu = document.querySelector('.submenu');
-
-        usernameDiv.addEventListener('click', function(event) {
-            event.preventDefault();
-            submenu.classList.toggle('active');
-        });
-    });
-
-    // 네비게이션 링크 클릭 시 'active' 클래스 토글
-    function toggleActive(event) {
-        var link = event.target; // 클릭한 링크 요소 가져오기
-        var links = document.querySelectorAll('#menu a'); // 모든 링크 요소 가져오기
-
-        // 모든 링크에서 'active' 클래스 제거
-        links.forEach(function(item) {
-            item.classList.remove('active');
-        });
-
-        // 클릭한 링크에 'active' 클래스 추가
-        link.classList.add('active');
-    }
-
-
-</script>
-</body>
-</html>
