--- src/main/resources/templates/employee/dashboard.html
@@ -1,145 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
-    <style>
-        .navbar-brand img {
-            width: 80%; /* 이미지의 원래 크기로 설정 */
-        }
-
-        #content-container {
-            display: flex;
-            margin-left: 300px;
-            margin-top: -150px;
-            /* content와 navbar 사이의 간격 조정 */
-        }
-
-        #content {
-            width: 80%; /* 화면 너비의 50%로 조정 */
-            height: 600px; /* 회색 배경의 높이 조정 */
-            background-color: lightgray;
-            padding: 20px; /* 내부 여백 추가 */
-            margin-left: 10%;
-            text-align: center;
-            white-space: nowrap;
-        }
-
-        .inner-box {
-            background-color: white;
-            padding: 20px;
-            margin-top: 20px;
-            height: 85%;
-        }
-
-        .search {
-            display: flex;
-            align-items: center;
-        }
-
-        input {
-            flex: 1; /* 입력 창이 가능한 최대 너비를 가짐 */
-            margin-right: 10px;
-        }
-
-        #img2 {
-            width: 17px;
-            margin-right: 10px;
-        }
-
-        #btn, #nextbtn {
-            font-size: 14px;
-            background-color: darkgray;
-            border-color: darkgray;
-            margin: 3px;
-        }
-
-        #viewBtn {
-
-            background-color: darkgray;
-        }
-        /* 각 네비게이션 항목 사이의 간격을 늘립니다. */
-        .navbar-nav .nav-item {
-            margin-right: 100px;
-        }
-        .form-label {
-            font-weight: bold;
-            text-align: right; /* 라벨을 오른쪽으로 정렬 */
-            padding-right: 10px; /* 오른쪽 여백 추가 */
-        }
-        nav{
-            white-space: nowrap;
-        }
-
-    </style>
-</head>
-<body>
-<div class="container-fluid">
-
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <div class="container">
-            <a class="navbar-brand" href="#"><img src="https://github.com/panghunJO/HitechProject1/assets/157236060/ade911fb-5b0b-409d-b99a-4b4d23ecff64" alt="Logo" ></a>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0" >
-                    <li class="nav-item">
-                        <a class="nav-link" href="/employee/repair/repair">작업관리</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="/employee/part/part">재고관리</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#">메일/채팅</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="employee/account/account">계정관리</a>
-                    </li>
-                </ul>
-                <p id="username" class="nav-link" th:text=|${userName}님|></p>
-                <ul class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                    </a>
-                    <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                        <li><a class="dropdown-item" href="#">마이페이지</a></li>
-                        <li><a class="dropdown-item" href="#">로그아웃</a></li>
-                    </ul>
-                </ul>
-            </div>
-        </div>
-    </nav>
-</div>
-
-</body>
-<!-- Bootstrap JS -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
-<script>
-    // 사용자 이름 클릭 시 서브메뉴 토글
-    document.addEventListener("DOMContentLoaded", function() {
-        var usernameDiv = document.getElementById('username');
-        var submenu = document.querySelector('.submenu');
-
-        usernameDiv.addEventListener('click', function(event) {
-            event.preventDefault();
-            submenu.classList.toggle('active');
-        });
-    });
-
-    // 네비게이션 링크 클릭 시 'active' 클래스 토글
-    function toggleActive(event) {
-        var link = event.target; // 클릭한 링크 요소 가져오기
-        var links = document.querySelectorAll('#menu a'); // 모든 링크 요소 가져오기
-
-        // 모든 링크에서 'active' 클래스 제거
-        links.forEach(function(item) {
-            item.classList.remove('active');
-        });
-
-        // 클릭한 링크에 'active' 클래스 추가
-        link.classList.add('active');
-    }
-</script>
-</html>
