--- src/main/resources/templates/employee/part/partAdd.html
@@ -1,177 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>PartAdd</title>
-    <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
-    <style>
-        .navbar-brand img {
-            width: 80%; /* 이미지의 원래 크기로 설정 */
-        }
-        #content-container {
-            display: flex;
-            margin-left: 300px;
-            margin-top: -150px;
-
-        }
-
-        #content {
-            width: 80%; /* 화면 너비의 50%로 조정 */
-            height: 600px; /* 회색 배경의 높이 조정 */
-            background-color: lightgray;
-            padding: 140px;
-            /* 내부 여백 추가 */
-            margin-left: 10%;
-            text-align: center;
-            white-space: nowrap;
-        }
-
-        .form-label {
-            font-weight: bold;
-            text-align: right; /* 라벨을 오른쪽으로 정렬 */
-            padding-right: 10px; /* 오른쪽 여백 추가 */
-        }
-
-        .form-control {
-            margin-bottom: 10px;
-        }
-
-        .btn-regist {
-            background-color: darkgray;
-            border-color: darkgray;
-
-
-        }
-
-        .navbar-nav .nav-item {
-            margin-right: 100px;
-        }
-        nav{
-            white-space: nowrap;
-        }
-
-
-    </style>
-</head>
-<body>
-<div class="container-fluid">
-
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <div class="container">
-            <a class="navbar-brand" href="#"><img src="https://github.com/panghunJO/HitechProject1/assets/157236060/ade911fb-5b0b-409d-b99a-4b4d23ecff64" alt="Logo" ></a>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0" >
-                    <li class="nav-item">
-                        <a class="nav-link" href="employee/repair/repair">작업관리</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="/employee/part/part">부품관리</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#">메일/채팅</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="/employee/account/account">계정관리</a>
-                    </li>
-                </ul>
-                <p id="username" class="nav-link" th:text=|${userName}님|></p>
-                <ul class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                    </a>
-                    <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                        <li><a class="dropdown-item" href="#">마이페이지</a></li>
-                        <li><a class="dropdown-item" href="localhost:8080/logout">로그아웃</a></li>
-                    </ul>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
-    <div class="container mt-5">
-        <label class="form-label" style="font-size: 20px;">부품관리</label><br><br>
-        <a class="nav-link" href="/employee/part/part">부품조회</a><br><br><br>
-    </div>
-
-    <div class="row">
-        <div class="col-md-12">
-            <div class="container" id="content-container">
-                <div class="container" id="content">
-
-                    <div class="container">
-                        <div class="row">
-                            <div class="col-md-9">
-                                <form action="/employee/part/partAdd" method="post">
-                                    <div class="row mb-3">
-<!--                                        <label for="partNumber" class="col-sm-4 col-form-label text-end">부품 번호</label>-->
-                                        <div class="col-sm-8">
-<!--                                            <input name="partCode" type="text" class="form-control" id="partNumber" >-->
-                                        </div>
-                                    </div>
-                                    <div class="row mb-3">
-                                        <label for="partName" class="col-sm-4 col-form-label text-end">부품 이름</label>
-                                        <div class="col-sm-8">
-                                            <input name="partName" type="text" class="form-control" id="partName" >
-                                        </div>
-                                    </div>
-                                    <div class="row mb-3">
-                                        <label for="partQuantity" class="col-sm-4 col-form-label text-end">부품 수량</label>
-                                        <div class="col-sm-8">
-                                            <input name="partstock" type="text" class="form-control" id="partQuantity" >
-                                        </div>
-                                    </div>
-                                    <div class="row mb-3">
-                                        <label for="partPrice" class="col-sm-4 col-form-label text-end">부품 가격</label>
-                                        <div class="col-sm-8">
-                                            <input name="partPrice" type="text" class="form-control" id="partPrice" >
-                                        </div>
-                                    </div>
-                                    <div class="row">
-                                        <div class="col text-end">
-                                            <button type="submit" class="btn btn-regist">등록</button>
-                                        </div>
-                                    </div>
-                                </form>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<!-- Bootstrap JS -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
-<script>
-    // 사용자 이름 클릭 시 서브메뉴 토글
-    document.addEventListener("DOMContentLoaded", function() {
-        var usernameDiv = document.getElementById('username');
-        var submenu = document.querySelector('.submenu');
-
-        usernameDiv.addEventListener('click', function(event) {
-            event.preventDefault();
-            submenu.classList.toggle('active');
-        });
-    });
-
-    // 네비게이션 링크 클릭 시 'active' 클래스 토글
-    function toggleActive(event) {
-        var link = event.target; // 클릭한 링크 요소 가져오기
-        var links = document.querySelectorAll('#menu a'); // 모든 링크 요소 가져오기
-
-        // 모든 링크에서 'active' 클래스 제거
-        links.forEach(function(item) {
-            item.classList.remove('active');
-        });
-
-        // 클릭한 링크에 'active' 클래스 추가
-        link.classList.add('active');
-    }
-</script>
-</body>
-</html>
